+++ packages/adapter-interfaces/src/types/Application.ts
@@ -3,38 +3,22 @@ import { z } from "@hono/zod-openapi";
 export const applicationInsertSchema = z.object({
   id: z.string(),
   name: z.string(),
-  callbacks: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default("")
-    .openapi({
-      description:
-        "Comma-separated list of URLs whitelisted to use as a callback to the client after authentication.",
-    }),
-  allowed_origins: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default("")
-    .openapi({
-      description:
-        "Comma-separated list of URLs allowed to make requests from JavaScript to Auth0 API (typically used with CORS). By default, all your callback URLs will be allowed. This field allows you to enter other origins if necessary. You can also use wildcards at the subdomain level. Query strings and hash information are not taken into account when validating these URLs.",
-    }),
-  web_origins: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default("")
-    .openapi({
-      description:
-        "Comma-separated list of allowed origins for use with Cross-Origin Authentication, Device Flow, and web message response mode.",
-    }),
-  allowed_logout_urls: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default("")
-    .openapi({
-      description:
-        "Comma-separated list of URLs that are valid to redirect to after logout from Auth0. Wildcards are allowed for subdomains.",
-    }),
+  callbacks: z.array(z.string()).openapi({
+    description:
+      "Comma-separated list of URLs whitelisted to use as a callback to the client after authentication.",
+  }),
+  allowed_origins: z.array(z.string()).openapi({
+    description:
+      "Comma-separated list of URLs allowed to make requests from JavaScript to Auth0 API (typically used with CORS). By default, all your callback URLs will be allowed. This field allows you to enter other origins if necessary. You can also use wildcards at the subdomain level. Query strings and hash information are not taken into account when validating these URLs.",
+  }),
+  web_origins: z.array(z.string()).openapi({
+    description:
+      "Comma-separated list of allowed origins for use with Cross-Origin Authentication, Device Flow, and web message response mode.",
+  }),
+  allowed_logout_urls: z.array(z.string()).openapi({
+    description:
+      "Comma-separated list of URLs that are valid to redirect to after logout from Auth0. Wildcards are allowed for subdomains.",
+  }),
   addons: z
     .record(z.string(), z.record(z.string(), z.union([z.string(), z.number()])))
     .optional()
@@ -42,16 +26,6 @@ export const applicationInsertSchema = z.object({
       description:
         "Addons associated with the client. The key is the addon's package name and the value is an object with the configuration for the addon.",
     }),
-  // @deprecated. Renamed to match the auth0 API
-  allowed_web_origins: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default(""),
-  allowed_callback_urls: z
-    .string()
-    .transform((val) => (val === null ? "" : val))
-    .default(""),
-
   email_validation: z
     .enum(["enabled", "disabled", "enforced"])
     .default("enforced")