+++ packages/kysely/src/applications/create.ts
@@ -2,15 +2,6 @@ import { Kysely } from "kysely";
 import { Application, ApplicationInsert } from "@authhero/adapter-interfaces";
 import { Database } from "../db";
 
-function toJsonString(value: string) {
-  return JSON.stringify(
-    value
-      .split(",")
-      .map((item) => item.trim())
-      .filter((item) => item.length),
-  );
-}
-
 export function create(db: Kysely<Database>) {
   return async (
     tenant_id: string,
@@ -20,15 +11,12 @@ export function create(db: Kysely<Database>) {
       created_at: new Date().toISOString(),
       updated_at: new Date().toISOString(),
       ...params,
-      // TODO: remove fallbacks
-      allowed_web_origins: params.allowed_origins,
-      allowed_callback_urls: params.callbacks,
-      allowed_origins: params.allowed_origins,
     };
 
-    const allowed_origins = toJsonString(params.allowed_origins);
-    const allowed_callback_urls = toJsonString(params.callbacks);
-    const callbacks = toJsonString(params.callbacks);
+    const allowed_origins = JSON.stringify(params.allowed_origins);
+    const callbacks = JSON.stringify(params.callbacks);
+    const web_origins = JSON.stringify(params.web_origins);
+    const allowed_logout_urls = JSON.stringify(params.allowed_logout_urls);
 
     await db
       .insertInto("applications")
@@ -38,9 +26,9 @@ export function create(db: Kysely<Database>) {
         disable_sign_ups: params.disable_sign_ups ? 1 : 0,
         addons: params.addons ? JSON.stringify(params.addons) : "{}",
         callbacks,
-        allowed_web_origins: allowed_origins,
-        allowed_callback_urls,
         allowed_origins,
+        web_origins,
+        allowed_logout_urls,
       })
       .execute();