+++ packages/kysely/src/applications/get.ts
@@ -23,11 +23,10 @@ export function get(db: Kysely<Database>) {
       ...application,
       disable_sign_ups: !!application.disable_sign_ups,
       addons: JSON.parse(application.addons),
-      // TODO: add callbacks
       callback: JSON.parse(application.callbacks),
       allowed_origins: JSON.parse(application.allowed_origins),
-      allowed_callback_urls: JSON.parse(application.allowed_callback_urls),
-      allowed_web_origins: JSON.parse(application.allowed_web_origins),
+      web_origins: JSON.parse(application.web_origins),
+      allowed_logout_urls: JSON.parse(application.allowed_logout_urls),
     });
   };
 }