+++ packages/kysely/src/applications/list.ts
@@ -13,6 +13,14 @@ export function list(db: Kysely<Database>) {
       ...result,
       disable_sign_ups: !!result.disable_sign_ups,
       addons: result.addons ? JSON.parse(result.addons) : {},
+      callbacks: result.callbacks ? JSON.parse(result.callbacks) : [],
+      allowed_origins: result.allowed_origins
+        ? JSON.parse(result.allowed_origins)
+        : [],
+      web_origins: result.web_origins ? JSON.parse(result.web_origins) : [],
+      allowed_logout_urls: result.allowed_logout_urls
+        ? JSON.parse(result.allowed_logout_urls)
+        : [],
     }));
 
     return {